Crate bmw_err

The BMW Error crate

This crate includes the main error structs, enums and macros used in BMW for building and mapping errors. This crate offers wrappers around the rust failure crate. The crate::map_err macro can be used to conveniently map errors from 3rd party crates into crate::ErrorKind in this crate. The crate::err macro can be used to generate errors. In most cases errors should be created using one of these two macros.

Examples

 // Example of the err macro
 use bmw_err::{Error, ErrorKind, ErrKind, err, map_err, try_into};
 use std::path::PathBuf;
 use std::fs::File;
 use std::io::Write;

 fn process_file(path: &str) -> Result<(), Error> {
     if ! PathBuf::from(path).exists() {
         return Err(err!(ErrKind::IllegalArgument, "path does not exist"));
     }

     // .. process file

     Ok(())
 }

 // Example of the map_err macro
 fn show_map_err() -> Result<(), Error> {
     // map the file open error to a 'Log' Error. The text of the original error will be
     // included in the mapped error.
     let mut x = map_err!(File::open("/invalid/log/path.log"), ErrKind::Log)?;
     x.write(b"test")?;

     // optionally an additional message can be included as below. The original
     // error's message will still be displayed.
     let file = map_err!(
         File::open("/path/to/something"),
         ErrKind::IO,
         "file open failed"
     )?;
     println!("file_type={:?}", file.metadata()?.file_type());

     Ok(())
 }

 // Example of the err macro
 fn return_err() -> Result<(), Error> {
     // return an error with the ErrKind::Misc with the text below
     Err(err!(ErrKind::Misc, "misc error occured"))
 }

 // Example of the try_into macro
 fn try_into_ex() -> Result<(), Error> {
     // try to convert a u128 to a usize
     let _v: usize = try_into!(100u128)?;
     Ok(())
 }

 fn main() -> Result<(), Error> {
     try_into_ex()?;
     assert!(return_err().is_err());
     assert!(show_map_err().is_err());
     let mut buf = PathBuf::new();
     buf.push("/path/to/err");
     assert!(process_file(buf.to_str().unwrap()).is_err());
     Ok(())
 }

Formatting

The crate::err macro may be used to format output just like std::println or std::format. See the example below for details.

 use bmw_err::{Error, ErrKind, err};

 // The error macro can be used like the format macro to format output
 fn test() -> Result<(), Error> {
     let  code = 404;
     // return an error with formatting
     Err(err!(ErrKind::Http, "The webserver returned the error code: {}", code))
 }

 fn main() -> Result<(), Error> {
     assert!(test().is_err());
     Ok(())
 }

Macros

cbreak

Macro to do a conditional break

err

Build the specified crate::ErrorKind and convert it into an crate::Error. The desired crate::ErrorKind is specified using the crate::ErrKind name enum.

map_err

Map the specified error into the crate::ErrKind enum name from this crate. Optionally specify an additional message to be included in the error.

try_into

Macro to map the try_from error into an appropriate error.

Structs

Error

Base Error struct which is used throughout BMW.

Enums

ErrKind

The kinds of errors in this crate. This enum is used to map to error names using the crate::err and crate::map_err macros.

ErrorKind

Kinds of errors that can occur.