Crate bmw_err

This crate includes the main error structs, enums and macros used in bmw for building and mapping errors. This crate offers wrappers around the rust failure crate. The crate::map_err macro can be used to conveniently map errors from 3rd party crates into crate::ErrorKind in this crate. The crate::err macro can be used to generate errors. In most cases errors should be created using one of these two macros.

Examples

 // Example of the err macro
 use bmw_err::{Error, ErrorKind, ErrKind, err, map_err};
 use std::path::PathBuf;
 use std::fs::File;
 use std::io::Write;

 fn process_file(path: &str) -> Result<(), Error> {
     if ! PathBuf::from(path).exists() {
         return Err(err!(ErrKind::IllegalArgument, "path does not exist"));
     }

     // .. process file

     Ok(())
 }

 // Example of the map_err macro
 fn show_map_err(do_error: bool) -> Result<(), Error> {
     // map the file open error to a 'Log' Error. The text of the original error will be
     // included in the mapped error.
     let mut x = map_err!(File::open("/invalid/log/path.log"), ErrKind::Log)?;
     x.write(b"test")?;

     // optionally an additional message can be included as below. The original
     // error's message will still be displayed.
     let file = map_err!(
         File::open("/path/to/something"),
         ErrKind::IO,
         "file open failed"
     )?;
     println!("file_type={:?}", file.metadata()?.file_type());

     Ok(())
 }

Macros

err

Build the specified crate::ErrorKind and convert it into an crate::Error. The desired crate::ErrorKind is specified using the crate::ErrKind name enum.

map_err

Map the specified error into the crate::ErrKind enum name from this crate. Optionally specify an additional message to be included in the error.

try_into

Macro to map the try_from error into an appropriate error.

Structs

Error

Base Error struct which is used throughout bmw.

Enums

ErrKind

The names of ErrorKinds in this crate. This enum is used to map to error names using the crate::err and crate::map_err macros.

ErrorKind

Kinds of errors that can occur.